add dynamic nav
