+++ src/components/header.js
@@ -6,13 +6,22 @@ import * as actions from '../actions';
 class Header extends Component {
 	renderLinks() {
 		if (this.props.authenticated) {
-			return (
-				<li className='nav-item'>
+			return [
+				<li className='nav-item' key={1}>
 					<div className='nav-link' onClick={() => this.props.signoutUser()}>
 						Sign Out
 					</div>
+				</li>,
+				<li className='nav-item' key={2}>
+					<Link className='nav-link' to='/profile'>Profile</Link>
+				</li>,
+				<li className='nav-item' key={3}>
+					<Link className='nav-link' to='/saved'>Saved</Link>
+				</li>,
+				<li className='nav-item' key={4}>
+					<Link className='nav-link' to='/messages'>Messages</Link>
 				</li>
-			);
+			];
 		} else {
 			return [
 				<li className='nav-item' key={1}>
@@ -26,10 +35,18 @@ class Header extends Component {
 		}
 	}
 
+	renderLogo() {
+		if (this.props.authenticated) {
+			return (<Link to='/home' className='navbar-brand'>Cards</Link>);
+		} else {
+			return (<Link to='/' className='navbar-brand'>Cards</Link>);
+		}
+	}
+
 	render() {
 		return (
 			<nav className='navbar navbar-light'>
-				<Link to='/' className='navbar-brand'>Cards</Link>
+				{this.renderLogo()}
 				<ul className='nav navbar-nav'>
 					{this.renderLinks()}
 				</ul>
