+++ src/screens/messages.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Messages extends Component {
+	render() {
+	    return (
+	    	<div>
+	      		This is the messages screen
+	      	</div>
+	    );
+	}
+}
+
+export default Messages;
