+++ src/screens/profile.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Profile extends Component {
+	render() {
+	    return (
+	    	<div>
+	      		This is the profile screen
+	      	</div>
+	    );
+	}
+}
+
+export default Profile;
