+++ src/screens/signup.js
@@ -1,13 +1,14 @@
 import React, { Component } from 'react';
 import {reduxForm, Field} from 'redux-form';
-import formInput from '../forms/formInput';
+import formInput from '../components/forms/formInput';
 import { connect } from 'react-redux';
-import * as actions from '../../actions';
+import * as actions from '../actions';
 
-class Signin extends Component {
-	handleFormSubmit({email, password}) {
+class Signup extends Component {
+	handleFormSubmit(formProps) {
+		// Call action creator to sign up the user
 		const { history } = this.props;
-		this.props.signinUser({ email, password }, history);
+		this.props.signupUser(formProps, history)
 	}
 
 	renderAlert() {
@@ -21,7 +22,7 @@ class Signin extends Component {
 	}
 
 	render() {
-		const {handleSubmit} = this.props;
+		const { handleSubmit } = this.props;
 
 		return (
 			<div>
@@ -40,9 +41,16 @@ class Signin extends Component {
 							   type='password'
 		                       component={formInput}/>
 					</fieldset>
+					<fieldset className='form-group'>
+						<label>Confirm Password:</label>
+						<Field name='passwordConfirm'
+							   placeholder='confirm password'
+							   type='password'
+		                       component={formInput}/>
+					</fieldset>
 					{this.renderAlert()}
 					<button action='submit' className='btn btn-primary'>
-						Sign in
+						Sign up
 					</button>
 				</form>
 				<a href='http://localhost:3090/auth/facebook' 
@@ -54,12 +62,35 @@ class Signin extends Component {
 	}
 }
 
-Signin = reduxForm({
-    form: 'signin'
-})(Signin);
+function validate(formProps) {
+	const errors = {};
 
-function mapStateToProps(state) {
-	return { errorMessage: state.auth.error };
+	if (!formProps.email) {
+		errors.email = 'Please enter an email';
+	}
+
+	if (!formProps.password) {
+		errors.password = 'Please enter a password';
+	}
+
+	if (!formProps.passwordConfirm) {
+		errors.passwordConfirm = 'Please enter a password confirmation';
+	}
+
+	if (formProps.password !== formProps.passwordConfirm) {
+		errors.password = 'Passwords must match';
+	}
+
+	return errors;
 }
 
-export default connect(mapStateToProps, actions)(Signin);
+Signup = reduxForm({
+    form: 'signup',
+    validate: validate
+})(Signup);
+
+function mapStateToProps(state) {
+	return { errorMessage: state.auth.error }
+}
+
+export default connect(mapStateToProps, actions)(Signup);
