+++ src/screens/welcome.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Welcome extends Component {
+	render() {
+	    return (
+	    	<div>
+	      		Welcome to Cards
+	      	</div>
+	    );
+	}
+}
+
+export default Welcome
